Implemented saving of custom goal size (after pinch gesture); implemented continuous priority to support

+++ Falco/GroupGoal.swift
@@ -17,7 +17,7 @@ struct GroupGoal: Goal {
 
     var name: String
     var details: String
-    var priority: Int
+    var priority: Double
     var endTime: NSDate
 
     var id: String { return _id }
@@ -64,7 +64,7 @@ struct GroupGoal: Goal {
 
     // MARK: Init
 
-    init(id: String = NSUUID().UUIDString, groupId: String, name: String, details: String, priority: Int = 0, endTime: NSDate, userCompletionTimes: [User: NSDate] = [:]) {
+    init(id: String = NSUUID().UUIDString, groupId: String, name: String, details: String, priority: Double = 0, endTime: NSDate, userCompletionTimes: [User: NSDate] = [:]) {
         self._id = id
         self._groupId = groupId
         self.name = name
@@ -77,7 +77,7 @@ struct GroupGoal: Goal {
     init(id: String, groupId: String, goalData: [String: AnyObject]) {
         let name = goalData[Constants.nameKey]! as! String
         let details = goalData[Constants.detailsKey]! as! String
-        let priority = goalData[Constants.priorityKey]! as! Int
+        let priority = goalData[Constants.priorityKey]! as! Double
         let endTime = NSDate(timeIntervalSince1970: NSTimeInterval(goalData[Constants.endTimeKey] as! NSNumber))
         let userData = goalData[Constants.userCompletionTimesKey]! as! [String: NSNumber]
         var userCompletionTimes: [User: NSDate] = [:]
