+++ Falco/PersonalGoal.swift
@@ -16,7 +16,7 @@ struct PersonalGoal: Goal {
 
     var name: String
     var details: String
-    var priority: Int
+    var priority: Double
     var endTime: NSDate
 
     var id: String { return _id }
@@ -33,7 +33,7 @@ struct PersonalGoal: Goal {
 
     // MARK: Init
 
-    init(id: String = NSUUID().UUIDString, name: String, details: String, priority: Int = 0, endTime: NSDate, completionTime: NSDate = Constants.incompleteTimeValue) {
+    init(id: String = NSUUID().UUIDString, name: String, details: String, priority: Double = 0, endTime: NSDate, completionTime: NSDate = Constants.incompleteTimeValue) {
         self._id = id
         self.name = name
         self.details = details
@@ -45,7 +45,7 @@ struct PersonalGoal: Goal {
     init(id: String, goalData: [String: AnyObject]) {
         let name = goalData[Constants.nameKey]! as! String
         let details = goalData[Constants.detailsKey]! as! String
-        let priority = goalData[Constants.priorityKey]! as! Int
+        let priority = goalData[Constants.priorityKey]! as! Double
         let endTime = NSDate(timeIntervalSince1970: NSTimeInterval(goalData[Constants.endTimeKey] as! NSNumber))
         let completionTime = NSDate(timeIntervalSince1970: NSTimeInterval(goalData[Constants.completionTimeKey] as! NSNumber))
         self.init(id: id, name: name, details: details, priority: priority, endTime: endTime, completionTime: completionTime)
